+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   > so it can be configured by env
 * stats: replace cpu usage with JDK built-in OperatingSystemMXBean.getProcessCpuLoad()
   > since jdk 17, java is aware of container, refer to https://developers.redhat.com/articles/2022/04/19/java-17-whats-new-openjdks-container-awareness
+* rate-limit: update rate control config api to make it more intuitive
+  > e.g. add("group", 10, 5, Duration.ofSeconds(1)) keeps 10 permits at most, fills 5 permits every second
+  > e.g. add("group", 20, 10, Duration.ofMinutes(5)) keeps 20 permits at most, fills 10 permits every 5 minutes
+* search: added retryOnConflict on UpdateRequest to handle
 
 ### 7.10.6 (04/19/2022 - 04/28/2022)